WIP: Enable index scan on leaf partitions directly
The fundamental problem is that relcache translator does not report the
indices present on leaf partition when we directly query the leaf
partition.

While generating relation metadata,
`CTranslatorRelcacheToDXL::PdrgpmdidRelIndexes()` invoked by
'`Pmdrel()`bails out and does not report indices if the relation is a
leaf partition.

In this commit; we now fetch the index oids for leaf partitions. Now a
leaf partition can obtain indices from root or indices defined on it
directly. Hence we fetch indices defined on root and on the leaf
partition itself.

Now since we have indexoids in relation metadata for leaf; we generate
index metadata via `Pmdindex()`.

`Pmdindex()` has special handling for leaf partitions. Consider scenario
where we are querying the root partition directly. Then the indices
fetched for root is nothing but the indices on its leaf partitions.
Hence `Pmdindex()` checks if the rel on which index is defined is a leaf
then it sets the rel to root based on the reasoning just mentioned.

However since we now want to support indices on leaf partitions; we need
special handling in `Pmdindex()`.

ORCAs way of representing and fetching the partial indices needs a
redesign since the current implementation causes complexities in
implementing `indexscan on leaf partitions` feature. For instance,

``` DROP TABLE IF EXISTS foo; CREATE TABLE foo (id int, someval int)
DISTRIBUTED by (id) PARTITION BY RANGE(someval) (PARTITION p1 START (1)
END (1000) INCLUSIVE, DEFAULT PARTITION other);

CREATE INDEX idx_foo on foo_1_prt_p1 USING btree (someval);

INSERT INTO foo (id, someval) select i,i from generate_series(1,100000)
i;

explain select * from foo_1_prt_other where someval=2000; ```

With the current commit, we incorrectly pick up index scan using
`idx_foo` for `foo_1_prt_other`. Because in `Pmdindex()` we get the root
of `foo_1_prt_other`, which is `foo` and then we fetch the logical index
oids on root. The logical indices are returned by `gpdb::Plgidx()`.
Because of this whole design of partial indices; this function returns
`idx_foo` as index on `foo` and uses this index for `foo_1_prt_other`.

Hence we should first work on removing the partial indices (till we have
a better way of implementing it) such that the when we try to get
indices on root; we will have only homogenous indices (indices defined
on all parts).

This commit can be merged once we have changes for removing the partial
indices in place.

To verify; the previous attempt of removing partial indexes
(83a2f87010241c0f301d0d4613700450d010e250) has been cherry-picked and
this commit is applied on top of it. Now for the above query we do not
use incorrect index and go ahead with table scan; However if we query
`foo_1_prt_p1`; we now use index scan.